The 'netplay' program.
Code for creating network connection is in function
fd_net and put in extra file net.{h,c}.

netplay can use a function from a patched ALSA-lib.
Makefile can build it when given path to patched lib in
variable ALSANC.

Documentation is builtin, and Makefile adjusted to
build netplay and variants netplay_ALSANC and a static
version of the ALSANC variant.