Introduce --mmap option to playhrt
Don't copy aplay here which only uses the non-mmap compatiblity functions
of the ALSA library. As before we trigger the writing with the high res
timer instead of interrupts.

With --mmap it is not necessary to use the patched ALSA library.
First tests showed an improved sound with --mmap option.